메인페이지에서 상품 이미지와 정보 내려받기 구현 완료

+++ src/main/java/com/shop/domain/Product.java
@@ -9,8 +9,8 @@ import lombok.Data;
 public class Product {
 	 private Long productNo;
 	 private String name;
-	 private int price;
-	 private int salePrice;
+	 private Long price;
+	 private Long salePrice;
 	 private Long categoryId;
 	 private String description;
 	 private String useYn;
